Add template tests for memory, repeated-S and repeated-H
