+++ surface_sim/circuit_blocks/rot_surface_code_css.py
@@ -57,6 +57,8 @@ __all__ = [
     "qec_round_iterator",
     "qec_round_pipelined",
     "qec_round_pipelined_iterator",
+    "gate_to_iterator",
+    "gate_to_iterator_pipelined",
 ]
 
 
@@ -381,6 +383,21 @@ def qec_round_pipelined_iterator(
 
 gate_to_iterator = {
     "TICK": qec_round_iterator,
+    "I": idle_iterator,
+    "S": log_fold_trans_s_iterator,
+    "X": log_x_iterator,
+    "Z": log_z_iterator,
+    "CX": log_trans_cnot_iterator,
+    "CNOT": log_trans_cnot_iterator,
+    "R": init_qubits_z0_iterator,
+    "RZ": init_qubits_z0_iterator,
+    "RX": init_qubits_x0_iterator,
+    "M": log_meas_z_iterator,
+    "MZ": log_meas_z_iterator,
+    "MX": log_meas_x_iterator,
+}
+gate_to_iterator_pipelined = {
+    "TICK": qec_round_pipelined_iterator,
     "I": idle_iterator,
     "S": log_fold_trans_s_iterator,
     "X": log_x_iterator,
