+++ surface_sim/circuit_blocks/rot_surface_code_xzzx.py
@@ -44,6 +44,8 @@ __all__ = [
     "init_qubits_x1_iterator",
     "qec_round",
     "qec_round_iterator",
+    "gate_to_iterator",
+    "gate_to_iterator_pipelined",
 ]
 
 
@@ -634,6 +636,18 @@ def qec_round_google(
 
 gate_to_iterator = {
     "TICK": qec_round_iterator,
+    "I": idle_iterator,
+    "X": log_x_iterator,
+    "Z": log_z_iterator,
+    "R": init_qubits_z0_iterator,
+    "RZ": init_qubits_z0_iterator,
+    "RX": init_qubits_x0_iterator,
+    "M": log_meas_z_iterator,
+    "MZ": log_meas_z_iterator,
+    "MX": log_meas_x_iterator,
+}
+gate_to_iterator_pipelined = {
+    "TICK": qec_round_pipelined_iterator,
     "I": idle_iterator,
     "X": log_x_iterator,
     "Z": log_z_iterator,
